# docs/swagger.yaml
basePath: /
definitions:
  apierror.APIError:
    properties:
      code:
        type: integer
      sysMessage:
        type: string
      timestamp:
        type: string
      userMessage:
        type: string
    type: object
  domain.Group:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      receivers_count:
        type: integer
      updated_at:
        type: string
    type: object
  domain.Receiver:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  handlers.CreateGroupRequest:
    properties:
      description:
        example: Marketing subscribers list
        maxLength: 256
        type: string
      name:
        example: Marketing
        maxLength: 128
        type: string
    required:
    - description
    - name
    type: object
  handlers.CreateGroupResponse:
    properties:
      id:
        example: c790g02f8n90184b23qg
        type: string
    type: object
...
  handlers.GetGroupResponse:
    properties:
      group:
        $ref: '#/definitions/domain.Group'
    type: object
...
  handlers.ListGroupReceiversResponse:
    properties:
      receivers:
        items:
          $ref: '#/definitions/domain.Receiver'
        type: array
    type: object
  handlers.ListGroupsResponse:
    properties:
      groups:
        items:
          $ref: '#/definitions/domain.Group'
        type: array
    type: object
...
  handlers.SendGroupRequest:
    properties:
      body:
        example: Important team updates...
        maxLength: 2048
        type: string
      group_id:
        example: c790g02f8n90184b23qg
        type: string
      subject:
        example: Team Announcement
        maxLength: 256
        type: string
    required:
    - body
    - group_id
    - subject
    type: object
...
  handlers.UpdateGroupRequest:
    properties:
      description:
        example: Product announcements list
        maxLength: 256
        type: string
      name:
        example: Product Updates
        maxLength: 128
        type: string
    type: object
  handlers.UpdateGroupResponse:
    properties:
      id:
        example: c790g02f8n90184b23qg
        type: string
    type: object
...
host: localhost:2500
info:
  contact: {}
  description: Carrpigeon email distribution and template service API.
  title: Carrpigeon API
  version: "1.0"
paths:
  /groups:
    get:
      description: Get a paginated list of receiver groups
      parameters:
      - default: 10
        description: Number of items to return (default 10)
        in: query
        name: limit
        type: integer
      - default: 0
        description: Number of items to skip (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ListGroupsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apierror.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apierror.APIError'
      summary: List groups
      tags:
      - groups
    post:
      consumes:
      - application/json
      description: Creates a new receiver group
      parameters:
      - description: Group creation data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateGroupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.CreateGroupResponse'
        "400":
          description: Bad Request or validation error
          schema:
            $ref: '#/definitions/apierror.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apierror.APIError'
      summary: Create group
      tags:
      - groups
  /groups/{id}:
    delete:
      description: Deletes a receiver group by ID
      parameters:
      - description: Group ID (xid)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apierror.APIError'
        "404":
          description: Group Not Found
          schema:
            $ref: '#/definitions/apierror.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apierror.APIError'
      summary: Delete group
      tags:
      - groups
    get:
      description: Get details of a receiver group by ID
      parameters:
      - description: Group ID (xid)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.GetGroupResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apierror.APIError'
        "404":
          description: Group Not Found
          schema:
            $ref: '#/definitions/apierror.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apierror.APIError'
      summary: Get group by ID
      tags:
      - groups
    put:
      consumes:
      - application/json
      description: Updates an existing receiver group
      parameters:
      - description: Group ID (xid)
        in: path
        name: id
        required: true
        type: string
      - description: Group update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateGroupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.UpdateGroupResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apierror.APIError'
        "404":
          description: Group Not Found
          schema:
            $ref: '#/definitions/apierror.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apierror.APIError'
      summary: Update group
      tags:
      - groups
  /groups/{id}/receivers:
    get:
      description: Get a paginated list of receivers belonging to a specific group
      parameters:
      - description: Group ID (xid)
        in: path
        name: id
        required: true
        type: string
      - default: 10
        description: Number of items to return (default 10)
        in: query
        name: limit
        type: integer
      - default: 0
        description: Number of items to skip (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ListGroupReceiversResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apierror.APIError'
        "404":
          description: Group Not Found
          schema:
            $ref: '#/definitions/apierror.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apierror.APIError'
      summary: List receivers in group
      tags:
      - groups
  /groups/{id}/receivers/{receiverId}:
    delete:
      description: Removes a receiver from a group
      parameters:
      - description: Group ID (xid)
        in: path
        name: id
        required: true
        type: string
      - description: Receiver ID (xid)
        in: path
        name: receiverId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apierror.APIError'
        "404":
          description: Group or receiver relation not found
          schema:
            $ref: '#/definitions/apierror.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apierror.APIError'
      summary: Remove receiver from group
      tags:
      - groups
    put:
      description: Associates an existing receiver with a group
      parameters:
      - description: Group ID (xid)
        in: path
        name: id
        required: true
        type: string
      - description: Receiver ID (xid)
        in: path
        name: receiverId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apierror.APIError'
        "404":
          description: Group or receiver not found
          schema:
            $ref: '#/definitions/apierror.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apierror.APIError'
      summary: Add receiver to group
      tags:
      - groups
...
  /send/group:
    post:
      consumes:
      - application/json
      description: Queues an email to all active receivers belonging to a specified
        group
      parameters:
      - description: Group email send payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.SendGroupRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
        "400":
          description: Bad Request or validation error
          schema:
            $ref: '#/definitions/apierror.APIError'
        "404":
          description: Group not found
          schema:
            $ref: '#/definitions/apierror.APIError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/apierror.APIError'
      summary: Send email to group
      tags:
      - send
...
swagger: "2.0"